refactor: rename components and update related IDs for consistency
