+++ app/frontend/components/AppFooterComponent.vue
