+++ app/frontend/components/AppHeaderComponent.vue
@@ -4,7 +4,7 @@
     class="z-50 flex w-full flex-wrap py-7 md:flex-nowrap md:justify-start"
   >
     <nav
-      class="relative mx-auto flex w-full max-w-7xl basis-full flex-wrap items-center px-4 md:grid md:grid-cols-12 md:px-6 md:px-8"
+      class="relative mx-auto flex w-full max-w-7xl basis-full flex-wrap items-center px-4 md:grid md:grid-cols-12 md:px-8"
       aria-label="Global"
     >
       <div class="md:col-span-3">
