+++ app/frontend/components/IndexPage.vue
