+++ app/frontend/components/SpotPickerComponent.vue
