+++ app/frontend/components/SpotShowComponent.vue
