+++ app/frontend/components/SpotsSearchPage.vue
