+++ app/views/index/show.html.erb
@@ -1,3 +1,3 @@
-<div id="app-header"></div>
-<div id="index"></div>
-<div id="app-footer"></div>
+<div id="app-header-component"></div>
+<div id="index-page"></div>
+<div id="app-footer-component"></div>
