+++ app/views/spots/new.html.erb
@@ -53,7 +53,7 @@
       </div>
 
       <div
-        id="spot-map-picker"
+        id="spot-picker-component"
         data-google-maps-api-key="<%= @google_maps_api_key %>"
         class="mt-4 h-96"
       ></div>
