+++ app/views/spots/search.html.erb
@@ -1,6 +1,6 @@
 <div class="container mx-auto px-4 py-8">
   <div
-    id="map-search-app"
+    id="spots-search-page"
     data-google-maps-api-key="<%= @google_maps_api_key %>"
   ></div>
 </div>
