+++ app/views/spots/show.html.erb
@@ -57,7 +57,7 @@
 
     <div>
       <div
-        id="spot-map"
+        id="spot-show-component"
         data-google-maps-api-key="<%= @google_maps_api_key %>"
         data-latitude="<%= @spot.latitude %>"
         data-longitude="<%= @spot.longitude %>"
